added projects model and updated version_control model
